+++ apps/docs/components/ExampleConfirmButton.tsx
@@ -1,14 +1,17 @@
+'use client';
+
 import { ConfirmButton } from '@enterwell/react-ui';
 
 export function ExampleConfirmButton() {
     return (
         // @highlight-start
         <ConfirmButton
-            header={'Potvrdite brisanje'}
-            message='Jeste li sigurni da želite obrisati ovaj zapis?'
             color="error"
             variant='contained'
-            onConfirm={() => { }}>
+            onConfirm={() => { }}
+            header={'Potvrdite brisanje'}
+            message='Jeste li sigurni da želite obrisati ovaj zapis?'
+        >
             Obriši
         </ConfirmButton>
         // @highlight-end