+++ apps/docs/components/ExampleSearchHeader.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { SearchHeader } from '@enterwell/react-ui';
 
 export function ExampleSearchHeader() {
@@ -5,7 +7,8 @@ export function ExampleSearchHeader() {
         <SearchHeader
             variant='h4'
             onSubmit={(searchTerm) => console.log(searchTerm)}
-            placeholder='Search by term'>
+            placeholder='Search by term'
+        >
             Search
         </SearchHeader>
     )