+++ apps/docs/components/ExampleUseDataGrid.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { TypedExtendedGridColDef, useDataGrid } from '@enterwell/react-mui-hooks';
 import { DataGridPro } from '@mui/x-data-grid-pro';
 import { useEffect } from 'react';
@@ -40,12 +42,13 @@ export function ExampleUseDataGrid() {
     };
 
     const dataGrid = useDataGrid({
+        tableId: 'example-use-data-grid',
         columns,
         onPage: handleOnPage
     });
 
     useEffect(() => {
-        dataGrid.filterChanged();
+        dataGrid.refreshTable();
     }, []);
 
     return (