+++ apps/docs/components/ExampleUsePromiseWithPromiseObject.tsx
@@ -1,4 +1,6 @@
-import { useState } from 'react';
+'use client';
+
+import { useMemo } from 'react';
 import { usePromise } from '@enterwell/react-hooks';
 
 function getItems() {
@@ -11,7 +13,8 @@ function getItems() {
 
 export function ExampleUsePromiseWithPromiseObject() {
     // @highlight-start
-    const { item, isLoading, error } = usePromise(getItems);
+    const getItemsMemo = useMemo(() => getItems(), []);
+    const { item, isLoading, error } = usePromise(getItemsMemo);
     // @highlight-end
 
     return (