+++ apps/docs/content/react-ui/_meta.js
@@ -1,0 +1,4 @@
+export default {
+    "about": "About",
+    "getting-started": "Getting Started"
+}