+++ packages/react-ui/ItemAccordion/index.tsx
@@ -1,0 +1,4 @@
+export * from "./ItemAccordion";
+export * from "./ItemAccordionSummary";
+export * from "./ItemAccordionDetails";
+export * from "./ItemAccordionActions";